refactor: Dynamically import ActivateAccountFlow
